+++ packages/headless/src/test/mock-insight-state.ts
@@ -1,7 +1,15 @@
 import {getConfigurationInitialState} from '../features/configuration/configuration-state';
+import {getFacetSetInitialState} from '../features/facets/facet-set/facet-set-state';
 import {getInsightConfigurationInitialState} from '../features/insight-configuration/insight-configuration-state';
 import {getInsightInterfaceInitialState} from '../features/insight-interface/insight-interface-state';
+import {getInsightCaseContextSearchInitialState} from '../features/insight-search/insight-case-context-state';
+import {getPaginationInitialState} from '../features/pagination/pagination-state';
+import {getQuerySetInitialState} from '../features/query-set/query-set-state';
+import {getQuerySuggestSetInitialState} from '../features/query-suggest/query-suggest-state';
+import {getQueryInitialState} from '../features/query/query-state';
+import {getResultPreviewInitialState} from '../features/result-preview/result-preview-state';
 import {getSearchHubInitialState} from '../features/search-hub/search-hub-state';
+import {getSearchInitialState} from '../features/search/search-state';
 import {InsightAppState} from '../state/insight-app-state';
 
 export function buildMockInsightState(
@@ -10,9 +18,17 @@ export function buildMockInsightState(
   return {
     configuration: getConfigurationInitialState(),
     insightConfiguration: getInsightConfigurationInitialState(),
+    search: getSearchInitialState(),
     insightInterface: getInsightInterfaceInitialState(),
     searchHub: getSearchHubInitialState(),
     version: 'unit-testing-version',
+    insightCaseContext: getInsightCaseContextSearchInitialState(),
+    query: getQueryInitialState(),
+    facetSet: getFacetSetInitialState(),
+    pagination: getPaginationInitialState(),
+    resultPreview: getResultPreviewInitialState(),
+    querySet: getQuerySetInitialState(),
+    querySuggest: getQuerySuggestSetInitialState(),
     ...config,
   };
 }